+++ spec/fixtures/14_stubs/in.rb
@@ -1,0 +1,5 @@
+class BananaTest < ActiveSupport::TestCase
+  test "random stuff from mocha" do
+    Banana.stubs(:delicious?).returns(true)
+  end
+end